Initial commit
Grid Tensor is a heavily C++ templated implementation of "AI Tensors". These tensors are more
about storing data and typical vector and matrix operations rather than a mathematical or
physical definition of a Tensor.

The implementation relies heavily on C++ templates and C++20 features. This provides and
abstraction layer for models that can be mostly agnostic to the underlying implementations
with optimizations for CPUs and accelerators.

Signed-off-by: Chris Zankel <chris@zankel.net>