tensor: change TensorBase to a non-templated struct
The idea of having those template arguments was for the base class to
implement common methods across all specific implementations by
providing all necessary template arguments to declare a Tensor for the
derived implementation.

However, most methods in the tensors are very unique, and if desired,
another class could be introduced for this.

Having a simpler base class makes using concepts easier.

This patch introduces a concept for a Tensor by requiring that any class
to be derived from TensorBase.

Signed-off-by: Chris Zankel <chris@zankel.net>