slowcpu: fix strides, use bytes as the smallest unit
Strides describe the increments (in bytes) to go between two consecutive
values in an 'axis'. Consecutive elements have the stride sizeof(TYPE).

For rank-3 tensors, the default strides would be [0, 1, 2]:
sizeof(TYPE) * dim[1] * dim[0], sizeof(TYPE) * dim[0], sizeof(TYPE)

There are two options, to use the 'TYPE' as the smallest unit, so
sizeof(TYPE) would be 1, or use bytes. The current implementation uses
bytes. Note that elements should still be aligned to the type.

Disable Add/Add for now, it requires shared buffers, and the current
implementation uses a temporary which is deleted before the addition
occurs.

Signed-off-by: Chris Zankel <chris@zankel.net>