tensor: use value_type instead of char for pointer arguments
There are different options for passing the pointer to the tensor
buffer, each comes with its benefits and problems:

 * value_type*  --> works better when data is type-aligned
 * char*        --> works better when data is not aligned
 * void*        --> doesn't really work with anything ;-)

This patch changes the pointer handling to value_type. This seems to
also align better to stl, where data() often returns also a pointer to
the first element and not a char* buffer.

Signed-off-by: Chris Zankel <chris@zankel.net>