tensor: add support for a tensor "view"
The TensorView provides a "view" into another tensor, which is typically
used to specify a sub-rank or sub-area of the tensor for copying data
in or out of a tensor.

The view is created by specifying the axes and additional offsets and
strides.

Views don't have ownership of the buffer, they merely map onto the
underlying buffer. When used as an lvalue, it always copies the data
from the rhs tensor (or view) into "its" tensor, regardless if the rhs
tensor is a rvalue-reference or not.

Signed-off-by: Chris Zankel <chris@zankel.net>