base: default template parameter breaks alias templates
One design goal was the ability to switch between tensor implementations
(for different devices) by using alias declarations:

   using Tensor = TENSOR<...>

(with the appropriate namespace and template arguments for TENSOR).

Using a default parameter breaks user-defined deduction rules:

  template <..., typename Allocator = std::allocator<T>>
  class Tensor {};

  template <...>
  Tensor(...) -> Tensor<... StaticAllocator>

However, the change drops the ability to use any custom allocator with
the default Tensor implementation. The idea was to use the default
tensor with a device-specific memory management/allocator. It's unclear
if this would have worked, though.

Signed-off-by: Chris Zankel <chris@zankel.net>