tensor/base: add copy/move constr./assign and clean up
Add the required copy/move constructors and assignment operators and use
Array in Tensor instead of manipulationg Array::data_ directly.

The copy constructor is actually deleted as Array doesn't know the
structure and would just blindly copy the entire buffer.

Also impelement an Realloc to delete the current buffer and allocate one
of the new size. The caller is responsible to copy data for a resize
function.

Signed-off-by: Chris Zankel <chris@zankel.net>