tensor: change stride from bytes to type
Align to C++ usages of strides, which defines is as the number of
elements instead of a byte size.

Using bytes also had limit use as there is a dange of unaligned access
unless adding a lot of additional code.

One use case could have been RGBx where x is unused, but that could be
handled as a different type.

Signed-off-by: Chris Zankel <chris@zankel.net>