tensor: refactor iterators to use index/indices instead of ptrs
Using indices instead of pointers makes it easier to use
"multi-dimensional" iterators in range operations.

The implementation is not optimal, though, especially when used
with operator++/--. This could be mediated by having specializations for
contiguous tensors, using a sentinel for the "end()" iterartor, etc.

Most arithmetic and other operators will likely not use these
increment/decrement operators, but should still be improved.

Signed-off-by: Chris Zankel <chris@zankel.net>