tensor/base: clean up folding and binary/unary implementation
Various changes and rearranging the code to reduce the functions:

Use a template argument to define contiguous elements instead of
different element numbers of dimension and strides. This reduces the
number of implementations from 4 to 2.

Don't pass strides through the "Fold" function to the operation, just
bypass them in the lambda.

Signed-off-by: Chris Zankel <chris@zankel.net>