tensor: major renaming of variables without functions change
Implement the following naming convention. There are many options, and
not one seems to be better than another.

 * use d for the destination tensor
 * use x,y,z,w for input tensors, w should be reserved for weights
 * strides are postfixed with the tensor: stride_x, stride_d
 * the iterator to the first element in a range is first_x, first_d

Note that the the implementation has the destination first, followed by
the inputs. This aligns with d = x + y, and older C memcpy, etc.
The new std::range and other newer functions, such as std::copy,
however, have the output last. Same for CUDA and Metal kernels. They
also often just use a, b, c for input and output.

So, a follow-up patch might change that order.

Signed-off-by: Chris Zankel <chris@zankel.net>