tensor: return array with const sizes for BroadcastStrides
When returning a std::array with the updated strides in
BroadcastStrides, return an array of "const ssize_t" instead of just
ssize_t. This can then be passed directly into a function requiring a
std::span of const ssize_t.

Passing const values between arrays and spans is very tricky, so this
might not be the final way it is implemented here.

Signed-off-by: Chris Zankel <chris@zankel.net>