tensor: rename TensorOperator to TensorOperation and file
While definitions are not always that clear, operators seem to be used
more for unary operations on function spaces (operator theory) or for
the symbol used by an operationr, for example the '+' in a + b.

This patch renames TensorOperator to TensorOperation. It provides a base
class for n-ary functions or operations.

Signed-off-by: Chris Zankel <chris@zankel.net>