tensor: refactor the Fold function to return folded strides [1/2]
The original version required the caller to analyze the dimensions and
contiguous flat to manage any resulting strides. The refactored version
will return the folded dimension and strides, broadcasts to the full
rank, if necessary (by adding '0' dimensions).

This first patch just renames the original Fold function to Fold2 (and
all references).

Signed-off-by: Chris Zankel <chris@zankel.net>