tensor: add a "Fold" function that does not broadcast strides
FoldBroadcast "broadcasts" strides aligning then to the same rank as
dimensions and filling it with "0". This is not useful for all
implementation. Provide a "Fold" function for those cases. The rank of
strides is reduced by the folded dimensions but can be less than the
folded dimension (or even empty for scalars).

Signed-off-by: Chris Zankel <chris@zankel.net>